Update Attribution file structure
