+++ context/app/static/js/components/Detail/Attribution/index.js
@@ -1,0 +1,3 @@
+import Attribution from './Attribution';
+
+export default Attribution;
